extract getElementsImages in lib/listings.ts to reuse it in EntityListingComment
